Move some files around, reduce RAM usage of skeleton/node map
