+++ TODO.txt
@@ -472,6 +472,8 @@ local gravity mode
 superfluids:
 - super high heat conductivity
 
+windows sometimes crashes loading saves (bad alloc)
+- rendering saves on vs build id:2532322
 
 debug element, stuff like a compass or clock dsplay
 
@@ -481,17 +483,25 @@ curved line tool
 
 brush size indicator
 
+options:
+- frame limiter
+- screenshot
+- record
+
 Add more reactions to existing elements:
 
 realistic gardening
 
 rope
 
-standardize code for fields
+standardize code for fields (refactor)
  - consistent casing (uppercase) for ufnctions
 - reduce circuit constructor size
 - standardize fields (caps and shit)
 - fix { or \n{
+- std::array
+- don't use vectors of vectors
+int length(not_null<Record*> p);
 
 generate wiki page py
 
@@ -550,10 +560,9 @@ iczero
 MOON and ICZR fight each other over the pause button
 IOVD - sorts particles 
 
-saving portals and wifi requires iterating everything - check if there IS wifi or portal before copying
+https://powdertoy.co.uk/Discussions/Thread/View.html?Thread=17513&PageNum=0#Message=252721
 
-faster renderer
-- experiment with multithreading the renderer
+saving portals and wifi requires iterating everything - check if there IS wifi or portal before copying
 
 BWBellairs element
 - ACU
@@ -587,6 +596,40 @@ make group filling a bit better maybe?
 
 TRFR : transformer
 
+circuits-refactor:
+- define types
+- conversions for - 2 and other node ID
+- get branch by node helper
+- make some checks like IS_GRND valuated at evaltime
+
+- 2nd node trimming
+    A - B - C -> A - C
+    - trim nodes in chips until there's only 1 (trim branches that are chip-chip)
+- Add node method
+- ReID() nodes method at end that rids of unused nodes
+- char with std::byte
+- setNodeId to set node id instead of immeutable and shit
+
+replace nodeId == 1 with a fancy #define
+
+limit number of nodes in a circuit anyways? no need to use short, maybe a byte will do
+optimize memory:
+- undo redo don't save circuit, recompute while paused
+
+ rename do floodfill
+
+- rename framework to someting reasonable
+- seperate file for Circuit class
+- consistent casing (uppercase) for ufnctions
+- reduce circuit constructor size
+- standardize fields (caps and shit)
+- fix { or \n{
+- std::array
+- don't use vectors of vectors
+int length(not_null<Record*> p);
+
+updating resistor values should trigger resolve
+
 check normal SPRK stuff 
 - like electrolysis
 - control stuff like PSTN, voltage = distnace to push?
@@ -614,6 +657,7 @@ else 0.7 V drop
 
 NPN: do opposite
 
+RSPRK removes circuit when deleted - move that to kill_part
 
 - magnetic field induces inductor current
 
@@ -622,6 +666,10 @@ realistic batteries
 op amp maybe
 
 pnp, npn transistors
+
+multi-node group:
+- are "dynamic"
+- constrain voltages at nodes
 
 photoresistor, potetiometer (user controlled?)
 LOGC - logic gate
