+++ src/simulation/circuits/format_circuits.h
@@ -1,7 +1,7 @@
 #ifndef FORMAT_CIRCUIT_H
 #define FOMRAT_CIRCUIT_H
 
-#include "common/String.h"
+class String;
 
 const char inc_prefixes[] = {'k', 'M', 'G', 'T', 'P', 'E', 'Z', 'Y'};
 const char dec_prefixes[] = {'m', 'u', 'n', 'p', 'f', 'a', 'z', 'y'};
